+++ pinot-core/src/main/java/org/apache/pinot/core/transport/QueryServer.java
@@ -32,9 +32,10 @@ import io.netty.handler.ssl.ClientAuth;
 import io.netty.handler.ssl.SslContextBuilder;
 import io.netty.handler.ssl.SslProvider;
 import java.util.concurrent.TimeUnit;
+import org.apache.pinot.common.config.TlsConfig;
 import org.apache.pinot.common.metrics.ServerMetrics;
+import org.apache.pinot.common.utils.TlsUtils;
 import org.apache.pinot.core.query.scheduler.QueryScheduler;
-import org.apache.pinot.core.util.TlsUtils;
 import org.apache.pinot.server.access.AccessControl;
 import org.apache.pinot.server.access.AllowAllAccessFactory;
